feat: added getTotalScore for Examination

+++ src/main/java/comp3111/examsystem/entity/Question.java
@@ -101,7 +101,7 @@
             this.type = type;
         }
 
-        public double getScore() {
+        public float getScore() {
             return score;
         }
